requirements.txt
beautifulsoup4==4.14.3
fastparquet==2025.12.0
pandas==3.0.1
pyyaml==6.0.3
requests==2.32.5
lxml==6.0.2
selenium==4.41.0
webdriver-manager==4.0.2